producer/requirements.txt
google-cloud-pubsub==2.19.0
requests==2.31.0
